good version

+++ app/src/main/java/ru/zuevs5115/deadlinedaemon/utils/TimeFormatter.kt
@@ -1,17 +1,15 @@
 package ru.zuevs5115.deadlinedaemon.utils
 
-import android.content.Context
-import ru.zuevs5115.deadlinedaemon.R
 import java.text.SimpleDateFormat
 import java.time.Duration
 import java.util.Date
 import java.util.Locale
 
-
+//just time formatter
 object TimeFormatter {
+    //convert interval to simple format
     fun formatNotificationInterval(seconds: Long): String {
-
-        val duration = Duration.ofSeconds(seconds.toLong())
+        val duration = Duration.ofSeconds(seconds)
         val days = duration.toDays()
         val hours = duration.toHours() % 24
         val minutes = duration.toMinutes() % 60
@@ -24,6 +22,7 @@ object TimeFormatter {
             if (secs > 0 && days == 0L && hours == 0L) append("$secs сек.")
         }.trim().ifEmpty { "0 сек." }
     }
+    //make date for last update status
     fun formatTimestamp(timestamp: Long): String {
         val format = SimpleDateFormat("dd MMM yyyy HH:mm:ss", Locale.getDefault())
         return format.format(Date(timestamp))
